PolySampleManifest added and SequenceManifest reverted to old
